Finally, softmax backward added - softmax backward was incorrect, added gradients according to https://towardsdatascience.com/derivative-of-the-softmax-function-and-the-categorical-cross-entropy-loss-ffceefc081d1 - tensor type will be float - conv nets will be tested next, finally

+++ neograd/autograd/utils.py
@@ -15,7 +15,7 @@ def process_data(data):
     if not isinstance(data, np.ndarray):
       data = np.array(data)
     try:
-      data = data.astype(np.float32)
+      data = data.astype(float)
     except ValueError:
       raise TypeError("Elements of data should be of type float or be typecastable to float")
   else:
